requirements.txt
langchain_core
python-dotenv
streamlit
langchain_community
langserve
fastapi
uvicorn
sse_starlette
bs4
pypdf
chromadb
faiss-cpu
sentence-transformers
langchain-huggingface
arxiv
wikipedia
groq
langchain-groq
pygithub